api/serializers.py
from django.contrib.auth.models import User
from rest_framework import serializers
from .models import *


# Serializers define the API representation.
class UserInfoSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserInfo
        fields = '__all__'


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'username', 'email', 'is_staff']
