requirements.txt
django>=3.0,<4.0
psycopg2
django-rest-framework
markdown
django-filter
django-guardian
django-rest-auth
uwsgi
django-cors-headers
Pillow
